requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.1.31
cffi==1.17.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
sentry-sdk==2.25.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
